更新自增字段并发用例
git-svn-id: http://192.168.20.11/sequoiadb/trunk@37555 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ testcase_new/story/java/src/test/java/com/sequoiadb/autoIncrement/Insert15958.java
@@ -34,7 +34,7 @@ public class Insert15958 extends SdbTestBase{
     private String indexName = "idIndex";
     private int threadNum = 10;
     private int threadInsertNum = 10000;
-    private int currentValue = 1;
+    private long currentValue = 1;
     
 	@BeforeClass
     public void setUp() {
